Refactoring
* Created shared features folder

* Moved log_in and sign_up to a spearate feature ('authentication')
