+++ lib/core/presentation/routes.dart
@@ -3,10 +3,10 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:picspeak/core/presentation/injectable/injectable_init.dart';
 import 'package:picspeak/core/presentation/splash_screen.dart';
-import 'package:picspeak/features/app_state/presentation/auth_forms/log_in_form/cubit/log_in_form_cubit.dart';
-import 'package:picspeak/features/app_state/presentation/auth_forms/log_in_form/log_in_form_screen.dart';
-import 'package:picspeak/features/app_state/presentation/auth_forms/sign_up_form/cubit/sign_up_form_cubit.dart';
-import 'package:picspeak/features/app_state/presentation/auth_forms/sign_up_form/sign_up_form_screen.dart';
+import 'package:picspeak/features/authentication/presentation/log_in_form/cubit/log_in_form_cubit.dart';
+import 'package:picspeak/features/authentication/presentation/log_in_form/log_in_form_screen.dart';
+import 'package:picspeak/features/authentication/presentation/sign_up_form/cubit/sign_up_form_cubit.dart';
+import 'package:picspeak/features/authentication/presentation/sign_up_form/sign_up_form_screen.dart';
 import 'package:picspeak/features/channels_browsing/presentation/channel_list_screen/cubit/channel_list_cubit.dart';
 import 'package:picspeak/features/channels_browsing/presentation/channel_screen/channel_screen.dart';
 import 'package:picspeak/features/channels_browsing/presentation/channel_list_screen/channel_list_screen.dart';
