+++ lib/features/authentication/presentation/log_in_form/log_in_form_screen.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
-import 'package:picspeak/features/app_state/presentation/auth_forms/sign_up_form/widgets/sign_up_fields_widget.dart';
+import 'package:picspeak/features/authentication/presentation/log_in_form/widgets/log_in_fields_widget.dart';
 
-class SignUpFormScreen extends StatelessWidget {
-  const SignUpFormScreen({Key? key}) : super(key: key);
+class LogInFormScreen extends StatelessWidget {
+  const LogInFormScreen({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -47,7 +47,7 @@ class SignUpFormScreen extends StatelessWidget {
                 child: Center(
                   child: Container(
                     height: 450,
-                    width: size.width > 600 ? 600 : size.width,
+                    width: size.width > 500 ? 500 : size.width,
                     decoration: BoxDecoration(
                       color: theme.colorScheme.background,
                       borderRadius: const BorderRadius.only(
@@ -60,7 +60,7 @@ class SignUpFormScreen extends StatelessWidget {
                         vertical: 15,
                         horizontal: 15,
                       ),
-                      child: SignUpFieldsWidget(),
+                      child: LogInFieldsWidget(),
                     ),
                   ),
                 ),
@@ -72,4 +72,3 @@ class SignUpFormScreen extends StatelessWidget {
     );
   }
 }
-
