+++ lib/features/authentication/presentation/sign_up_form/cubit/sign_up_form_cubit.dart
@@ -1,8 +1,8 @@
 import 'package:bloc/bloc.dart';
 import 'package:injectable/injectable.dart';
-import 'package:picspeak/features/app_state/domain/interactors/sign_up_interactor.dart';
+import 'package:picspeak/features/authentication/domain/interactors/sign_up_interactor.dart';
 import 'package:picspeak/features/app_state/presentation/app_state/blocs/app_state_cubit.dart';
-import 'package:picspeak/features/app_state/presentation/auth_forms/sign_up_form/cubit/sign_up_form_data.dart';
+import 'package:picspeak/features/authentication/presentation/sign_up_form/cubit/sign_up_form_data.dart';
 
 @Injectable()
 class SignUpFormCubit extends Cubit<SignUpFormData> {
