+++ lib/features/authentication/presentation/widgets/main_auth_button.dart
