+++ lib/features/channels_browsing/data/channel_repository_impl.dart
@@ -2,12 +2,12 @@ import 'package:flutter/cupertino.dart';
 import 'package:injectable/injectable.dart';
 import 'package:picspeak/core/utils/result/result.dart';
 import 'package:picspeak/core/utils/utils.dart';
-import 'package:picspeak/features/channels_browsing/domain/entities/channel/channel.dart';
-import 'package:picspeak/features/channels_browsing/domain/entities/event/event.dart';
-import 'package:picspeak/features/channels_browsing/domain/entities/event_recording_entry/event_recording_entry.dart';
-import 'package:picspeak/features/channels_browsing/domain/entities/talk/talk.dart';
-import 'package:picspeak/features/channels_browsing/domain/entities/talk_recording_entry/talk_recording_entry.dart';
-import 'package:picspeak/features/channels_browsing/domain/entities/user/user.dart';
+import 'package:picspeak/features/shared/domain/entities/channel/channel.dart';
+import 'package:picspeak/features/shared/domain/entities/event/event.dart';
+import 'package:picspeak/features/shared/domain/entities/event_recording_entry/event_recording_entry.dart';
+import 'package:picspeak/features/shared/domain/entities/talk/talk.dart';
+import 'package:picspeak/features/shared/domain/entities/talk_recording_entry/talk_recording_entry.dart';
+import 'package:picspeak/features/shared/domain/entities/user/user.dart';
 import 'package:picspeak/features/channels_browsing/domain/repositories/channel_repository.dart';
 
 @LazySingleton(as: ChannelRepository)
