+++ lib/features/channels_browsing/domain/get_channel_interactor.dart
@@ -1,7 +1,7 @@
 import 'package:injectable/injectable.dart';
 import 'package:picspeak/core/domain/interactors/await_for_connection_interactor.dart';
 import 'package:picspeak/core/utils/result/result.dart';
-import 'package:picspeak/features/channels_browsing/domain/entities/channel/channel.dart';
+import 'package:picspeak/features/shared/domain/entities/channel/channel.dart';
 import 'package:picspeak/features/channels_browsing/domain/repositories/channel_repository.dart';
 
 @LazySingleton()
