+++ lib/features/channels_browsing/domain/get_channel_recordings_interactor.dart
@@ -1,6 +1,6 @@
 import 'package:injectable/injectable.dart';
 import 'package:picspeak/core/utils/result/result.dart';
-import 'package:picspeak/features/channels_browsing/domain/entities/event_recording_entry/event_recording_entry.dart';
+import 'package:picspeak/features/shared/domain/entities/event_recording_entry/event_recording_entry.dart';
 import 'package:picspeak/features/channels_browsing/domain/repositories/channel_repository.dart';
 
 @LazySingleton()
