+++ lib/features/channels_browsing/domain/repositories/channel_list_repository.dart
@@ -1,4 +1,4 @@
-import 'package:picspeak/features/channels_browsing/domain/entities/channel/channel.dart';
+import 'package:picspeak/features/shared/domain/entities/channel/channel.dart';
 import 'package:picspeak/core/utils/result/result.dart';
 
 abstract class ChannelListRepository {
