+++ lib/features/channels_browsing/domain/repositories/channel_repository.dart
@@ -1,7 +1,7 @@
 import 'package:injectable/injectable.dart';
 import 'package:picspeak/core/utils/result/result.dart';
-import 'package:picspeak/features/channels_browsing/domain/entities/channel/channel.dart';
-import 'package:picspeak/features/channels_browsing/domain/entities/event_recording_entry/event_recording_entry.dart';
+import 'package:picspeak/features/shared/domain/entities/channel/channel.dart';
+import 'package:picspeak/features/shared/domain/entities/event_recording_entry/event_recording_entry.dart';
 
 abstract class ChannelRepository {
   Future<Result<Channel>> loadChannel(String channelId);
