+++ lib/features/channels_browsing/presentation/channel_screen/cubit/channel_recordings_cubit.dart
@@ -1,7 +1,7 @@
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:injectable/injectable.dart';
 import 'package:picspeak/core/utils/loading_state/loading_state.dart';
-import 'package:picspeak/features/channels_browsing/domain/entities/event_recording_entry/event_recording_entry.dart';
+import 'package:picspeak/features/shared/domain/entities/event_recording_entry/event_recording_entry.dart';
 import 'package:picspeak/features/channels_browsing/domain/get_channel_recordings_interactor.dart';
 
 @Injectable()
