+++ lib/features/channels_browsing/presentation/channel_screen/pages/channel_home/channel_home_page.dart
@@ -2,7 +2,7 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:picspeak/core/utils/loading_state/loading_state.dart';
-import 'package:picspeak/features/channels_browsing/domain/entities/channel/channel.dart';
+import 'package:picspeak/features/shared/domain/entities/channel/channel.dart';
 import 'package:picspeak/features/channels_browsing/presentation/channel_screen/cubit/channel_cubit.dart';
 import 'package:picspeak/features/channels_browsing/presentation/channel_screen/pages/channel_home/channel_home_page_loading.dart';
 import 'package:picspeak/features/channels_browsing/presentation/channel_screen/pages/channel_home/channel_home_page_ready.dart';
