+++ lib/features/channels_browsing/presentation/channel_screen/pages/channel_info/channel_info_ready.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:picspeak/features/channels_browsing/domain/entities/channel/channel.dart';
+import 'package:picspeak/features/shared/domain/entities/channel/channel.dart';
 
 class ChannelInfoReady extends StatelessWidget {
   const ChannelInfoReady({
