+++ lib/features/channels_browsing/presentation/channel_screen/pages/channel_recordings/widgets/talk_recording_entry_widget.dart
@@ -2,7 +2,7 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:picspeak/core/presentation/widgets/overlayed_ink_well.dart';
 import 'package:picspeak/core/utils/utils.dart';
-import 'package:picspeak/features/channels_browsing/domain/entities/talk_recording_entry/talk_recording_entry.dart';
+import 'package:picspeak/features/shared/domain/entities/talk_recording_entry/talk_recording_entry.dart';
 
 class TalkRecordingEntryWidget extends StatelessWidget {
   const TalkRecordingEntryWidget({
