+++ lib/features/channels_browsing/presentation/channel_screen/pages/channel_recordings/widgets/talk_recordings_list.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:picspeak/features/channels_browsing/domain/entities/talk_recording_entry/talk_recording_entry.dart';
+import 'package:picspeak/features/shared/domain/entities/talk_recording_entry/talk_recording_entry.dart';
 import 'package:picspeak/features/channels_browsing/presentation/channel_screen/pages/channel_recordings/widgets/talk_recording_entry_widget.dart';
 
 class TalkRecordingsList extends StatelessWidget {
