+++ lib/features/channels_browsing/presentation/event_details_screen/widgets/talk_entry_widget.dart
@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:picspeak/core/presentation/widgets/my_card.dart';
 import 'package:picspeak/core/presentation/widgets/overlayed_ink_well.dart';
 import 'package:picspeak/core/utils/utils.dart';
-import 'package:picspeak/features/channels_browsing/domain/entities/talk/talk.dart';
+import 'package:picspeak/features/shared/domain/entities/talk/talk.dart';
 
 class TalkEntryWidget extends StatelessWidget {
   const TalkEntryWidget({
