+++ lib/features/shared/domain/entities/channel/channel.freezed.dart
