+++ lib/features/shared/domain/entities/event/event.freezed.dart
