+++ lib/features/shared/domain/entities/event_recording_entry/event_recording_entry.dart
@@ -1,6 +1,6 @@
 import 'package:freezed_annotation/freezed_annotation.dart';
-import 'package:picspeak/features/channels_browsing/domain/entities/event/event.dart';
-import 'package:picspeak/features/channels_browsing/domain/entities/talk_recording_entry/talk_recording_entry.dart';
+import 'package:picspeak/features/shared/domain/entities/event/event.dart';
+import 'package:picspeak/features/shared/domain/entities/talk_recording_entry/talk_recording_entry.dart';
 part 'event_recording_entry.freezed.dart';
 
 @freezed
