+++ lib/features/shared/domain/entities/event_recording_entry/event_recording_entry.freezed.dart
