+++ lib/features/shared/domain/entities/talk/talk.dart
