+++ lib/features/shared/domain/entities/talk/talk.freezed.dart
