+++ lib/features/shared/domain/entities/talk_recording_entry/talk_recording_entry.dart
@@ -1,5 +1,5 @@
 import 'package:freezed_annotation/freezed_annotation.dart';
-import 'package:picspeak/features/channels_browsing/domain/entities/talk/talk.dart';
+import 'package:picspeak/features/shared/domain/entities/talk/talk.dart';
 part 'talk_recording_entry.freezed.dart';
 
 @freezed
