+++ lib/features/shared/domain/entities/user/user.freezed.dart
